FUUUUCK
# Conflicts:
#	code/modules/clothing/rogueclothes/armor.dm
#	code/modules/clothing/rogueclothes/gloves.dm
#	code/modules/clothing/rogueclothes/hats.dm

+++ code/modules/roguetown/roguemachine/merchant/goldface.dm
@@ -100,6 +100,10 @@
 				playsound(loc, 'sound/misc/gold_misc.ogg', 100, FALSE, -1)
 				update_icon()
 				return attack_hand(user)
+	if(istype(P, /obj/item/roguecoin/aalloy))
+		return
+	if(istype(P, /obj/item/roguecoin/inqcoin))	
+		return			
 	if(istype(P, /obj/item/roguecoin))
 		budget += P.get_real_price()
 		qdel(P)
